+++ src/core.cpp
@@ -4,4 +4,4 @@
 #include <pkgctl/pkgctl.h>
 
 static_assert(pkgctl::version_major == 0);
-static_assert(pkgctl::version_minor == 30);
+static_assert(pkgctl::version_minor == 31);