+++ tests/run-direct.sh
@@ -10,7 +10,7 @@ trap 'rm -rf "$tmp"' EXIT HUP INT TERM
 
 core_modules='libcrypto libpkgsource libpkgcatalog libpkgcatalog-acquire libpkgstate libpkgstate-posix libpkgstate-plan libpkgstate-apply libpkgfetch libpkgbuild libpkgbuild-exec libpkgbuild-image libpkgsource-plan libpkgbuild-plan libpkgimage libpkgplan libpkgexec libpkgapply libpkgapply-posix libpkgapply-exec libpkgresolve libpkgtransaction libpkgcheck libpkgcheck-exec'
 pkg-config --exists \
-  'libpkgbuild-exec >= 2.1.0' 'libpkgbuild-exec < 3.0.0' \
+  'libpkgbuild-exec >= 2.2.0' 'libpkgbuild-exec < 3.0.0' \
   'libpkgcheck-exec >= 0.4.0' 'libpkgcheck-exec < 1.0.0'
 core_cflags=$(pkg-config --cflags $core_modules)
 core_libs=$(pkg-config --libs $core_modules)